# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handlers.CreateNoteRequest:
    properties:
      content:
        type: string
      conversation_id:
        type: string
    required:
    - content
    - conversation_id
    type: object
...
  handlers.MigrationResponse:
    properties:
      message:
        type: string
      migrated_count:
        type: integer
    type: object
...
  models.Channel:
    properties:
      config:
        type: string
      created_at:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      qr_code:
        type: string
      qr_expires_at:
        type: string
      session:
        description: session identifier for WhatsApp integration
        type: string
      status:
        description: disconnected, connecting, connected
        type: string
      tenant_id:
        type: string
      type:
        description: whatsapp, webchat, etc.
        type: string
      updated_at:
        type: string
      webhook_token:
        type: string
      webhook_url:
        type: string
    required:
    - name
    - session
    - type
    type: object
  models.ChannelListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/models.SwaggerChannelWithConversationCount'
        type: array
      page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
  models.Conversation:
    properties:
      ai_enabled:
        type: boolean
      assigned_agent:
        $ref: '#/definitions/models.User'
      assigned_agent_id:
        type: string
      channel:
        $ref: '#/definitions/models.Channel'
      channel_id:
        type: string
      created_at:
        type: string
      customer:
        allOf:
        - $ref: '#/definitions/models.Customer'
        description: Relations
      customer_id:
        type: string
      id:
        type: string
      is_archived:
        type: boolean
      is_pinned:
        type: boolean
      last_message_at:
        type: string
      priority:
        type: string
      status:
        description: open, closed, waiting
        type: string
      tenant_id:
        type: string
      unread_count:
        type: integer
      updated_at:
        type: string
    type: object
...
  models.Customer:
    properties:
      birth_date:
        type: string
      created_at:
        type: string
      document:
        description: CPF/CNPJ
        type: string
      email:
        type: string
      gender:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      notes:
        type: string
      phone:
        description: Only numbers, no formatting
        type: string
      tenant_id:
        type: string
      updated_at:
        type: string
    required:
    - phone
    type: object
...
  models.Message:
    properties:
      content:
        type: string
      conversation:
        allOf:
        - $ref: '#/definitions/models.Conversation'
        description: Relations
      conversation_id:
        type: string
      created_at:
        type: string
      customer:
        $ref: '#/definitions/models.Customer'
      customer_id:
        type: string
      delivered_at:
        type: string
      direction:
        description: in, out
        type: string
      external_id:
        type: string
      filename:
        type: string
      forwarded_from:
        type: string
      id:
        type: string
      is_note:
        description: true for internal notes, false for regular messages
        type: boolean
      is_read:
        type: boolean
      media_type:
        type: string
      media_url:
        type: string
      metadata:
        type: string
      read_at:
        type: string
      reply_to_id:
        type: string
      routine_id:
        description: reference to routine that generated this message
        type: string
      source:
        description: whatsapp, chat
        type: string
      status:
        description: sent, delivered, read, failed
        type: string
      tenant_id:
        type: string
      type:
        description: text, image, audio, video, document
        type: string
      updated_at:
        type: string
      user_id:
        description: null for incoming messages
        type: string
      user_name:
        description: name of the user who sent the message
        type: string
      webhook_id:
        type: string
    type: object
...
  models.SwaggerChannelWithConversationCount:
    properties:
      access_token:
        type: string
      business_phone:
        type: string
      conversation_count:
        type: integer
      created_at:
        type: string
      id:
        type: string
      instance_id:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      phone_number:
        type: string
      tenant_id:
        type: string
      type:
        type: string
      updated_at:
        type: string
      webhook_token:
        type: string
      webhook_url:
        type: string
    type: object
...
  models.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      last_login_at:
        type: string
      name:
        type: string
      phone:
        type: string
      role:
        type: string
      tenant_id:
        description: null for system admins
        type: string
      updated_at:
        type: string
    required:
    - email
    - name
    - role
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Multi-tenant SaaS para vendas e atendimento via WhatsApp
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: IAFarma API
  version: "1.0"
paths:
...
  /channels:
    get:
      consumes:
      - application/json
      description: Get a list of channels with pagination
      parameters:
      - default: 20
        description: Limit
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ChannelListResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List channels
      tags:
      - channels
    post:
      consumes:
      - application/json
      description: Create a new channel
      parameters:
      - description: Channel data
        in: body
        name: channel
        required: true
        schema:
          $ref: '#/definitions/models.Channel'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Channel'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create channel
      tags:
      - channels
  /channels/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a channel by ID (only if it has no conversations)
      parameters:
      - description: Channel ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a channel
      tags:
      - channels
    get:
      consumes:
      - application/json
      description: Get a channel by its ID
      parameters:
      - description: Channel ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Channel'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get channel by ID
      tags:
      - channels
    put:
      consumes:
      - application/json
      description: Update an existing channel
      parameters:
      - description: Channel ID
        in: path
        name: id
        required: true
        type: string
      - description: Channel data
        in: body
        name: channel
        required: true
        schema:
          $ref: '#/definitions/models.Channel'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Channel'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update channel
      tags:
      - channels
  /channels/{id}/migrate-conversations:
    post:
      consumes:
      - application/json
      description: Migrate all conversations from source channel to destination channel
      parameters:
      - description: Source Channel ID
        in: path
        name: id
        required: true
        type: string
      - description: Migration request
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.MigrationResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Migrate conversations between channels
      tags:
      - channels
...
  /messages:
    get:
      consumes:
      - application/json
      description: Get list of messages for a specific conversation with pagination
      parameters:
      - description: Conversation ID
        in: query
        name: conversation_id
        required: true
        type: string
      - default: 50
        description: Limit
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Message'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List messages by conversation
      tags:
      - messages
    post:
      consumes:
      - application/json
      description: Create a new message
      parameters:
      - description: Message data
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/models.Message'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Message'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create message
      tags:
      - messages
  /messages/{id}:
    get:
      consumes:
      - application/json
      description: Get a message by its ID
      parameters:
      - description: Message ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Message'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get message by ID
      tags:
      - messages
  /messages/notes:
    post:
      consumes:
      - application/json
      description: Create an internal note for a conversation
      parameters:
      - description: Note data
        in: body
        name: note
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateNoteRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Message'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create internal note
      tags:
      - messages
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"